Add files via upload
This solution involves using ANTLR to generate java from a grammar that describes the parsing of the input(ten pin games).
the lexer converts input into tokens, the grammar specifies the expected arrival patterns for the tokens.
A visitor design pattern is used to operate on the grammar rules as they are encountered. This results in the data being stored in a model of the game and then at the end of the parse the calculation is called.